# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/media-literacy/courses": {
      "get": {
        "operationId": "MediaLiteracyController_getCourses",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isPublished",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isFeatured",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ]
      },
      "post": {
        "operationId": "MediaLiteracyController_createCourse",
        "summary": "Create a new course (Admin/Moderator)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/media-literacy/courses/{slug}": {
      "get": {
        "operationId": "MediaLiteracyController_getCourse",
        "summary": "Get course details",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ]
      },
      "patch": {
        "operationId": "MediaLiteracyController_updateCourse",
        "summary": "Update a course (Admin/Moderator)",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "MediaLiteracyController_deleteCourse",
        "summary": "Delete a course (Admin only)",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/media-literacy/courses/{courseId}/enroll": {
      "post": {
        "operationId": "MediaLiteracyController_enrollInCourse",
        "summary": "Enroll in a course",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/media-literacy/courses/{courseId}/progress": {
      "get": {
        "operationId": "MediaLiteracyController_getCourseProgress",
        "summary": "Get course progress for current user",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/media-literacy/lessons/{lessonId}/complete": {
      "post": {
        "operationId": "MediaLiteracyController_markLessonComplete",
        "summary": "Mark lesson as complete/incomplete",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkLessonCompleteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/media-literacy/my-courses": {
      "get": {
        "operationId": "MediaLiteracyController_getMyCourses",
        "summary": "Get my enrolled courses",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/media-literacy/courses/{slug}/publish": {
      "patch": {
        "operationId": "MediaLiteracyController_toggleCoursePublish",
        "summary": "Toggle course publish status (Admin/Moderator)",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/media-literacy/courses/{courseSlug}/lessons": {
      "post": {
        "operationId": "MediaLiteracyController_createLesson",
        "summary": "Create a new lesson for a course (Admin/Moderator)",
        "parameters": [
          {
            "name": "courseSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/media-literacy/lessons/{lessonId}": {
      "patch": {
        "operationId": "MediaLiteracyController_updateLesson",
        "summary": "Update a lesson (Admin/Moderator)",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "MediaLiteracyController_deleteLesson",
        "summary": "Delete a lesson (Admin only)",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media Literacy"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/api/v1/faq": {
      "get": {
        "operationId": "FaqController_getAllFAQs",
        "summary": "Get all published FAQs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FAQ"
        ]
      },
      "post": {
        "operationId": "FaqController_createFAQ",
        "summary": "Create a new FAQ (Admin/Moderator)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFaqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "FAQ"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/faq/{id}": {
      "get": {
        "operationId": "FaqController_getFAQ",
        "summary": "Get FAQ by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FAQ"
        ]
      },
      "patch": {
        "operationId": "FaqController_updateFAQ",
        "summary": "Update an FAQ (Admin/Moderator)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFaqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FAQ"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "FaqController_deleteFAQ",
        "summary": "Delete an FAQ (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FAQ"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/faq/{id}/publish": {
      "patch": {
        "operationId": "FaqController_toggleFAQPublish",
        "summary": "Toggle FAQ publish status (Admin/Moderator)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FAQ"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Tahaqaq 360 API",
    "description": "Fact-checking and Media Literacy Platform API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "MarkLessonCompleteDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction to Media Literacy"
          },
          "slug": {
            "type": "string",
            "example": "introduction-to-media-literacy"
          },
          "description": {
            "type": "string",
            "example": "Learn the fundamentals of media literacy and critical thinking"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Beginner",
              "Intermediate",
              "Advanced"
            ],
            "example": "Beginner"
          },
          "duration": {
            "type": "number",
            "example": 480,
            "description": "Duration in minutes"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "learningObjectives": {
            "example": [
              "Understand media bias",
              "Identify misinformation",
              "Analyze sources critically"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prerequisites": {
            "example": [
              "Basic reading comprehension",
              "Internet access"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coverImage": {
            "type": "string",
            "example": "https://example.com/course-image.jpg"
          },
          "isPublished": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "title",
          "slug",
          "description",
          "difficulty",
          "duration",
          "order",
          "learningObjectives",
          "prerequisites",
          "coverImage",
          "isPublished"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction to Media Literacy"
          },
          "slug": {
            "type": "string",
            "example": "introduction-to-media-literacy"
          },
          "description": {
            "type": "string",
            "example": "Learn the fundamentals of media literacy and critical thinking"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Beginner",
              "Intermediate",
              "Advanced"
            ],
            "example": "Beginner"
          },
          "duration": {
            "type": "number",
            "example": 480,
            "description": "Duration in minutes"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "learningObjectives": {
            "example": [
              "Understand media bias",
              "Identify misinformation",
              "Analyze sources critically"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prerequisites": {
            "example": [
              "Basic reading comprehension",
              "Internet access"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coverImage": {
            "type": "string",
            "example": "https://example.com/course-image.jpg"
          },
          "isPublished": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "CreateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Understanding Confirmation Bias"
          },
          "content": {
            "type": "string",
            "example": "# Confirmation Bias\n\nConfirmation bias is..."
          },
          "duration": {
            "type": "number",
            "example": 15
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "videoUrl": {
            "type": "string",
            "example": "https://youtube.com/watch?v=example"
          },
          "resources": {
            "example": [
              {
                "type": "pdf",
                "title": "Lesson Notes",
                "url": "https://example.com/notes.pdf"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "content",
          "duration",
          "order",
          "videoUrl",
          "resources"
        ]
      },
      "UpdateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Understanding Confirmation Bias"
          },
          "content": {
            "type": "string",
            "example": "# Confirmation Bias\n\nConfirmation bias is..."
          },
          "duration": {
            "type": "number",
            "example": 15
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "videoUrl": {
            "type": "string",
            "example": "https://youtube.com/watch?v=example"
          },
          "resources": {
            "example": [
              {
                "type": "pdf",
                "title": "Lesson Notes",
                "url": "https://example.com/notes.pdf"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateFaqDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "example": "What is fact-checking?"
          },
          "answer": {
            "type": "string",
            "example": "Fact-checking is the process of verifying the accuracy of information..."
          },
          "category": {
            "type": "string",
            "example": "General"
          },
          "order": {
            "type": "number",
            "example": 1,
            "description": "Display order (lower = higher priority)"
          },
          "isPublished": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "question",
          "answer",
          "category",
          "order",
          "isPublished"
        ]
      },
      "UpdateFaqDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "example": "What is fact-checking?"
          },
          "answer": {
            "type": "string",
            "example": "Fact-checking is the process of verifying the accuracy of information..."
          },
          "category": {
            "type": "string",
            "example": "General"
          },
          "order": {
            "type": "number",
            "example": 1,
            "description": "Display order (lower = higher priority)"
          },
          "isPublished": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  }
}